Implementação de UserList com Listas personalizadas de users

+++ Repositories/Interfaces/IGameRepository.cs
@@ -6,9 +6,9 @@ namespace LudoVault.Repositories.Interfaces
     {
         public Task<GameModel> Criar(GameModel game);
         public Task<List<GameModel>> BuscarTodos();
-        public Task<GameModel> BuscarPorId(long id);
-        public Task<GameModel> Atualizar(GameModel game, long id);
-        public Task<bool> Deletar(long id);
-        public Task<List<GameRatingModel>> BuscarRatings(long id);
+        public Task<GameModel> BuscarPorId(int id);
+        public Task<GameModel> Atualizar(GameModel game, int id);
+        public Task<bool> Deletar(int id);
+        public Task<List<GameRatingModel>> BuscarRatings(int id);
     }
 }
