+++ Repositories/Interfaces/IPublisherRepository.cs
@@ -7,7 +7,7 @@ namespace LudoVault.Repositories.Interfaces
         public Task<PublisherModel> Criar(PublisherModel publisher);
         public Task<PublisherModel> Atualizar(PublisherModel publisher);
         public Task<List<PublisherModel>> BuscarTodos();
-        public Task<PublisherModel> BuscarPorId(long id);
-        public Task<bool> Excluir(long id);
+        public Task<PublisherModel> BuscarPorId(int id);
+        public Task<bool> Excluir(int id);
     }
 }
