+++ Services/Interfaces/IGameServices.cs
@@ -6,10 +6,10 @@ namespace LudoVault.Services.Interfaces
     public interface IGameServices
     {
         public Task<GameResponse> CriarGame(GameRequest game);
-        public Task<GameResponse> AtualizarGame(GameRequest game, long id);
+        public Task<GameResponse> AtualizarGame(GameRequest game, int id);
         public Task<List<GameResponse>> BuscarGames();
-        public Task<GameResponse> BuscarGamePorId(long id);
-        public Task<bool> RemoverGame(long id);
-        public Task<GameRatingListUsersResponse> BuscarRatingsPorIdGame(long id);
+        public Task<GameResponse> BuscarGamePorId(int id);
+        public Task<bool> RemoverGame(int id);
+        public Task<GameRatingListUsersResponse> BuscarRatingsPorIdGame(int id);
     }
 }
