+++ Services/Interfaces/IPublisherServices.cs
@@ -6,9 +6,9 @@ namespace LudoVault.Services.Interfaces
     public interface IPublisherServices
     {
         public Task<PublisherResponse> CriarPublisher(PublisherRequest publisher);
-        public Task<PublisherResponse> AtualizarPublisher(PublisherRequest publisher, long id);
+        public Task<PublisherResponse> AtualizarPublisher(PublisherRequest publisher, int id);
         public Task<List<PublisherResponse>> BuscarPublishers();
-        public Task<PublisherResponse> BuscarPublisherPorId(long id);
-        public Task<bool> ExcluirPublisher(long id);
+        public Task<PublisherResponse> BuscarPublisherPorId(int id);
+        public Task<bool> ExcluirPublisher(int id);
     }
 }
